major changes to oom_notifierd
oom_notifierd now works on a per-session basis.  It only notifies
the current tty when an OOM event occurs. Multiple oom_notifierd
instances can be listening in the same cgroup but each is in charge
of notifying the tty (or non-tty) that spawned it.  When the
controlling tty or non-tty that launched oom_notifierd exits, the
oom_notifierd instance will terminate within a specified timeout
window (currently 15 seconds).